pyproject.toml
[tool.poetry]
name = "linkedin-bias-auditor"
version = "0.1.0"
description = "LinkedIn Retrieval Bias Auditor"
authors = ["Antigravity <antigravity@google.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.9.0"
httpx = "^0.27.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.2.0"
sqlalchemy = "^2.0.0"
pyyaml = "^6.0"
numpy = "^1.26.0"
tqdm = "^4.66.0"
scipy = "^1.11.0"
statsmodels = "^0.14.0"
torch = {version = ">=2.0.0,<3.0.0", python = ">=3.11,<3.15"}
transformers = {version = ">=4.40.0", python = ">=3.11,<3.15"}
accelerate = {version = ">=0.24.0", python = ">=3.11,<3.15"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-httpx = "^0.30.0"
ruff = "^0.3.0"
mypy = "^1.8.0"
types-pyyaml = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

[tool.mypy]
strict = true
ignore_missing_imports = true
